requirements.txt
google-genai>=1.0.0
pandas
pytest
python-dotenv
streamlit
